ndctl, monitor: improve error reporting throughout monitor.c
In several places in the ndctl monitor, we were losing useful error
information (from 'errno' for example), and just returning a simple '1'
or '-1'. Fix these to capture and propagate the correct errors
everywhere.

In the case of notify_dimm_event(), don't error out for failures of
json_object_new_*. Follow the precedent of util/json.c and only add the
object to its parent if the 'new' function was successful.

Cc: QI Fuli <qi.fuli@jp.fujitsu.com>
Reviewed-by: QI Fuli <qi.fuli@jp.fujitsu.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>