ndctl inject-smart: add an option to uninject smart fields
The smart injection command in ndctl was missing an option to uninject
the different inject-able fields. The APIs in libndctl already allow for
this by setting the 'enable' flag to false. Add a *-uninject argument
for each of the injectable fields that sets this flag and disables
injected values. Also add an "uninject-all" option to uninject all
fields.

Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>