ndctl, inject-smart: continue in spite of errors for uninject-all
When the missing command submissions for ctrl-temperature were added to
inject-smart, as a side effect, uninject-all also started trying to
uninject ctrl-temperature. Since DSMs for smart injection for certain
platforms don't support ctrl-temperature injection, this failed the
command.

For the uninject-all command, add a sctx flag that allows command
submissions to continue in spite of errors. The error messages are still
printed out, so they are not lost. This way, uninject-all does a
best-effort uninjection of all fields, even if some fields are
unsupported by the given platform.

Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>