ndctl: Add 'list' verbose options
The informational and miscellaneous flag options are becoming more
numerous, and can be difficult to remember what can be listed. This
patch adds a 'verbose' option that increases the detail listed by
automatically enabling options and less essential information.

The verbose option can be repeated multiple times to increase the
detail. There are currently three levels of verbose with output detail
documented in the ndctl list manpage.

Suggested-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Keith Busch <keith.busch@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>