ndctl: release v63
This release incorporates functionality up to the 4.20 kernel, and a
number of bug fixes and improvements.

Highlights include switching the documentation build to asciidoctor by
default, fixes to destroy-namespace for reporting the number of
namespaces acted upon, using the namespace badblocks listings exported
by the kernel, and making them available to non-root users, a new helper
for retrieving the dirty-shutdown-count, reverting the udev rule to set
the shutdown count latch and cache the dirty-shutdown-count, and fixing
the ndctl-monitor daemon to exit successfully in the absence of NVDIMMs.

Commands:
    destroy-namespace: fix number of namespaces reported
    check-labels: fix the number of labels checked reporting
    monitor: exit daemon with success when no DIMMs found

Tests:
    Fix a missing include for list_smart_dimm
    pfn-meta-errors.sh: new test for clearing errors in the volatile
        'struct page' metadata area

APIs:
    ndctl_dimm_get_dirty_shutdown
    ndctl_namespace_get_first_badblock
    ndctl_namespace_get_next_badblock