ndctl, daxctl: Add type-safety to command harness
A recent attempt to drop casts from ndctl/monitor.c ran afoul of the
'struct btt_chk' usage of an embedded 'struct log_ctx'. It is
generically useful to be able to reuse the log functions and fragile to
force them to be placed at offset zero in the container context.
Instead, just teach the harness the difference between ndctl and daxctl
commands. Let the leaf command implementations receive their expected
context with the correct type.

Reported-by: Vishal Verma <vishal.l.verma@intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>