ndctl/monitor: Fix / cleanup log_file()
The log_file() helper opens the log file for each message. Avoid the
scenario where the fopen() fails and just open it at init time.

This usage is also broken because if the vasprintf() fails it will try
to print via log_file again via fail(), and potentially lead to infinite
recursion.

Cc: Qi Fuli <qi.fuli@jp.fujitsu.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>