libndctl: ignore ENXIO in ndctl_namespace_delete
The kernel will return ENXIO and fail attempts to set the size of a
namespace that has already been deleted. When we are deleting a
namespace, we don't care that it has been previously destroyed, or is
in an uninitialized state.

Ignore the ENXIO from setting the size to zero in this case, and report
to the callers of this API that the deletion was successful.

This addresses github issue #73.

Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>