ndctl, Documentation: Allow for Makefile variables in Documentation
We were starting to accumulate automake variables such as configuration
file paths that were also referenced in Documentation man pages.
However, until now, the man pages simply hard coded these paths to their
default values. If a distribution were to configure such a path to
something other than the default, the man pages would be out of sync
with the reality.

Arrange for Makefile variables to be piped into an 'attrs.adoc' target
(the variables to be piped in this manner still have to be listed
explicitly). The different asciidoc(tor) source files can then include
attrs.adoc to use these variables.

Finally, convert instances of '/etc/ndctl/monitor.conf' in the monitor
documentation to use this new facility.

Cc: QI Fuli <qi.fuli@jp.fujitsu.com>
Cc: Dave Jiang <dave.jiang@intel.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Acked-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>