ndctl: add passphrase management commands
Add three new commands: 'setup-passphrase', 'update-passphrase' and
'remove-passphrase'.

setup-passphrase allows the user to perform the initial security setup
by generating encrypted passphrases for the DIMMs, and adding them to
the keyring.

update-passphrase allows for changing the passphrase and optionally,
also changing the encryption key (kek) for the passphrase.

remove-passphrase removes the passphrase and the associated key
material.

The supporting infrastructure for these commands that adds all aspects
of key management is added in ndctl/util. Some of this might be
standardizable as libndctl APIs in the future, but until another customer
of these interfaces appears, it is left unexposed.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>
[vishal: collapse remove-passphrase into this]
[vishal: man page fixups]
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>