ndctl, rpmbuild: Add a dummy readme file to ndctl_keysdir
At build time, ndctl_keysdir (/etc/nddctl/keys/) was empty, and thus
it never gets created. Since the directory is listed in the %files
section of the RPM spec, RPM checks for it and fails since it is not
present.

Fix this by adding a simple 'keys.readme' file to the keys dorectory
that points the user to the setup-passphrase man page, and installing it
during the build.

Cc: Dan Williams <dan.j.williams@intel.com>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>