ndctl: allow a "zero-key" for secure erase
Allow a "zero-key", i.e. a key with all zeroes, to be passed to
ndctl-sanitize-dimm for the 'crypto-erase' option. A key may not be
required for crypto-erase, and this option enables 'resetting' the
security state in the case of a lost key.

The "zero-key" semantic was already implemented for master-passphrase
and overwrite, so for completeness make the "zero-key" capability
common across all security commands. With this in place it makes more
sense to support platforms that might allow "passphrase-less" security
operations.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>