libndctl: Implement the "smart_get_health" dimm-op for Hyper-V
Show the health info of the NVDIMM using the "Get Health Information"
DSM (Function Index 1) as defined in [1]

With this, "ndctl list --dimms --health" on an Hyper-V NVDIMM shows:

  {
    "dev":"nmem0",
    "id":"04d5-01-1701-00000000",
    "handle":0,
    "phys_id":0,
    "health":{
      "health_state":"ok"
    }
  }

If the NVDIMM has an error, the string "ok" is replaced with either
"fatal", "critical", or "non-critical".

If ndctl fails to retrieve the health info due to some reason, the
health_state is displayed as "unknown".

[1]: http://www.uefi.org/RFIC_LIST ("Virtual NVDIMM 0x1901").

Signed-off-by: Dexuan Cui <decui@microsoft.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>