ndctl, monitor: Don't require the support of ND_CMD_SMART_THRESHOLD
An NVDIMM that doesn't support ND_CMD_SMART_THRESHOLD, may still support
ND_CMD_SMART or a similar variant. Allow such an NVDIMM to be monitored
with "ndctl monitor".

For such an NVDIMM, while threshold based events (such as temperatures)
won't work, 'absolute' events (such as 'fatal health state') will be
monitor-able, and that is useful to have even in the absence of
threshold based monitoring.

Cc: QI Fuli <qi.fuli@jp.fujitsu.com>
Signed-off-by: Dexuan Cui <decui@microsoft.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>