ndctl: add a 'clear-errors' command
Add a new command, ndctl-clear-errors, to clear any errors (badblocks)
on a namespace. This is in preparation for a 'system-ram' mode for
devdax devices using the kernel's 'kmem' facility. Since the device is
being used as volatile RAM, we can take the opportunity to clear any
badblocks on the device before reconfiguration, so that the user doesn't
come across one unexpectedly.

Make this error clearing facility generic to all namespace types (i.e.
devdax, fsdax, and raw; sector mode namespaces are not supported). To
clear errors, use the "Clear Uncorrectable Errors" ACPI DSM command via
the helpers provided by libndctl.

Cc: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>