ndctl/namespace.c: fix an unchecked return value
Static analysis complains that we were neglecting to check the return
value from the ndctl_enable_namespace() call in nstype_clear_badblocks().

The instance pointed out is in an error-out path, so there isn't much to
do there, however if the enable fails, we can check for it and print a
message. This ends up being more user friendly in addition to quelling
the warning.

Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>