libndctl/inject: Refuse error injection for BTT namespaces
Error injection on a BTT namespace would treat the namespace as 'raw'
for the purposes of the injection. This can be useful for development,
but to a user this can be surprising, as injecting with --block=1 would
corrupt the BTT info block, and the BTT would be lost.

The unit tests do not rely on injecting errors directly into a BTT
namespace - they convert the namespace to 'raw' mode before performing
such an injection. For development and testing purposes, we will still
retain this ability by enforcing that the BTT namespace be explicitly
forced into raw mode before injection.

Reported-by: Marek de Rosier <marekx.de.rosier@intel.com>
Reported-by: Dan Williams <dan.j.williams@intel.com>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>