ndctl/monitor: make the daemon exit message 'info' level
The 'ndctl monitor daemon started' message is at an 'info' log level,
but the corresponding exit message (if no DIMMs are found for
monitoring) was only at a 'debug' level. As a result, on a system
without NVDIMMs, the sysadmin might see a 'monitor started' message, but
a monitor wouldn't actually be running. Making the exit message the same
log level ensures that no confusion arises out of this.

Link: https://github.com/pmem/ndctl/issues/102
Cc: QI Fuli <qi.fuli@jp.fujitsu.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Reported-by: Adam Borowski <kilobyte@angband.pl>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>