ndctl, check: Ensure mmap of BTT sections work with 64K page-sizes
On PPC64 which uses a 64K page-size, ndctl-check-namespace fails on a BTT
namespace with following error:

  # ndctl check-namespace namespace0.0 -vv
  namespace0.0: namespace_check: checking namespace0.0
  namespace0.0: btt_discover_arenas: found 1 BTT arena
  namespace0.0: btt_create_mappings: mmap arena[0].info [sz = 0x1000,
    off = 0x1000] failed: Invalid argument
  error checking namespaces: Invalid argument
  checked 0 namespaces

An error occurs when btt_create_mappings() tries to mmap the sections
of the BTT device which are usually 4K offset aligned. However the
mmap() syscall expects the 'offset' argument to be in multiples of
page-size, hence it returns EINVAL causing the btt_create_mappings()
to error out.

To fix this, wrap mmap(2) and munmap(2) with helpers that perform the
mmap to a system page boundary, and then adjust the returned pointers if
needed, and update btt_create_mappings() to use these.

Reported-by: Harish Sriram <harish@linux.ibm.com>
Signed-off-by: Vaibhav Jain <vaibhav@linux.ibm.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>