Documentation: refactor 'bus options' into its own include
Create a new 'xable-bus-options.txt' file that is included everywhere
that previously relied upon xable-region-options.txt for bus options
inclusion.

While at it, fix the formatting for the first description text block
from an include - the empty line at the start caused the first block to
be incorrectly indented. This fixes both the dimm and region variants.

Cc: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>